+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Ensure combobox scroll controller has a client attached before using it ([#620](https://github.com/bdlukaa/fluent_ui/issues/620))
 - Correctly use `TextFormBox.initialValue`
 - Added `TreeViewState.toggleItem`, which toggles the item expanded state ([#493](https://github.com/bdlukaa/fluent_ui/issues/493))
+- Ensure `NavigationView` pane items are brought into view when selected
 
 ## 4.1.0